requirements.txt
requests
gspread
oauth2client
pydantic
PyYAML
urllib3
